Replace PythonLinter with NewPythonLinter
